// weeb-dashboard/lib/plugins/myanimelist/types.d.ts
export interface MyAnimeListConfig {
    enabled: boolean;
    sections: string[];
    username: string;
    favorites_max?: number;
}
// ...
export interface BasicAnimeFavorite {
    mal_id: number;
    title: string;
    image: string;
    start_year: number;
    type: string;
}
export interface FullAnimeFavorite extends BasicAnimeFavorite {
    synopsis: string;
    score: number;
    popularity: number;
    episodes: number | null;
    status: string;
    rank: number;
    year: number | null;
    genres?: {
        name: string;
    }[];
}